Decouple core from FunctionInitializer
